Update index.html
GPS Accuracy Logging: I've added logging for the accuracy of the GPS data (position.coords.accuracy) in the showUserLocation function. You can use this information to understand if there is a significant error in the GPS data. Depending on the device, GPS accuracy can vary between a few meters to over a kilometer.

Node Snapping Update: The findClosestNode function now uses latLng.distanceTo() from Leaflet to calculate the distance between the clicked point and all available nodes, ensuring it properly finds the closest point based on the actual map click.